need to cast sizes to double otherwise division will result in 0

+++ cass/input/file_input.cpp
@@ -113,7 +113,10 @@ public:
    *
    * @return the current progress
    */
-  double progress() {return _file.tellg()/_filesize;}
+  double progress()
+  {
+    return static_cast<double>(_file.tellg()) / static_cast<double>(_filesize);
+  }
 
   /** retrieve the number of events processed by this thread
    *
